% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipping.R
\name{flipping}
\alias{flipping}
\title{Flipping an image horizontally in black and white }
\usage{
flipping(image, output_path)
}
\arguments{
\item{image}{path of input file}

\item{output_path}{string, path for the output image file}
}
\value{
an image file at the specified output path in .png format
}
\description{
Flipping an image in either horizontal or vertical direction
}
\examples{
flip("images/samples.jpg", "images/flip.png")
}
