% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding.R
\name{padding}
\alias{padding}
\title{Add a padding to the border of the image}
\usage{
padding(image, width)
}
\arguments{
\item{image}{path of input file}

\item{width}{the pixels of the padding, the padding width for left, right, top, bottom are the same}
}
\value{
an image matrix that has been padded
}
\description{
Add a padding to the border of the image
}
