% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotates an image about it's center by the given degree.}
\usage{
rotate(image, degree)
}
\arguments{
\item{image}{path of input file}

\item{degree}{int of degree}
}
\value{
an image with the specified degree rotated
}
\description{
Rotates an image about it's center by the given degree.
}
