% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipping.R
\name{flipping}
\alias{flipping}
\title{Flipping an image horizontally in Black & White}
\usage{
flipping(input_path, output_path)
}
\arguments{
\item{output_path}{string, path for the output image file}

\item{image}{path of input file}
}
\value{
an image file at the specified output path in .png format
}
\description{
Flipping an image horizontally in Black & White
}
\examples{
flipping("images/samples.png", "images/flip.png")
}
