% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grayscale.R
\name{grayscale}
\alias{grayscale}
\title{Converting a color image into a grayscale image.}
\usage{
grayscale(image)
}
\arguments{
\item{image}{string, path for the input image file}
}
\value{
an image file in gray scale
}
\description{
Converting a color image into a grayscale image.
}
\examples{
grayscale("images/samples.jpg")
}
